.history/requirements_20250119170159.txt
faiss-cpu==1.9.0.post1
faiss-gpu==1.7.2
huggingface-hub==0.27.1
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-huggingface==0.1.2
numpy==1.26.4
PyPDF2==3.0.1
# torch @ https://download.pytorch.org/whl/cu121_full/torch-2.5.1%2Bcu121-cp310-cp310-linux_x86_64.whl
transformers==4.47.1
